fix : RemittanceList.vue에서 모달 관련 코드 분리(RemittanceModal.vue, UserDepositModal.vue) 및 공용 CSS, 태그 전용 CSS 추가
