+++ frontend/src/views/transaction/style/Transaction.style.css
@@ -1,0 +1,38 @@
+/* 전체 스타일 정의 */
+
+.total-frame {
+  width: 98%;
+  margin: 10px;
+  padding: 20px;
+  background-color: #fff;
+  border-radius: 8px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+.title-bar {
+  display: flex;
+  align-items: center;
+  width: 90%;
+}
+
+.title {
+  font-size: 2em;
+  font-weight: bold;
+  flex: 1;
+}
+
+.icon-button-group {
+  display: flex;
+  margin-right: 50px;
+  gap: 30px;
+}
+
+.card {
+  width: 100%;
+}
+
+.title {
+  justify-content: center;
+  font-size: 1.5rem;
+  font-weight: bold;
+}
